move get_from_context to new file
