+++ services/traction/api/db/models/base.py
@@ -11,6 +11,7 @@ from sqlalchemy import Column, func, text, select, desc, String
 from sqlalchemy.sql.expression import Select
 from sqlalchemy.dialects.postgresql import UUID, TIMESTAMP, VARCHAR, ARRAY
 from sqlmodel.ext.asyncio.session import AsyncSession
+from api.endpoints.dependencies.tenant_context import get_from_context
 
 from api.db.session import async_session
 from api.endpoints.models.v1.errors import NotFoundError
@@ -27,13 +28,20 @@ class BaseSchema(pydantic.BaseModel):
 class BaseModel(SQLModel, BaseSchema):
     __mapper_args__ = {"eager_defaults": True}
 
+    def __init__(self):
+        self._logger = logging.getLogger(type(self).__name__)
+
     @classmethod
     def tenant_select(cls, fallback_tenant_id: UUID = None) -> Select:
-        logger.warning(
+        self.logger.warning(
             """tenant_select being called on table not using
             TenantScopedModel, making unsafe call"""
         )
         return select(cls)
+
+    @property
+    def logger(self):
+        return self._logger
 
     @classmethod
     async def update_by_id(cls, item_id: UUID, values: dict):
@@ -82,7 +90,6 @@ class TenantScopedModel(BaseModel):
 
     @classmethod
     def tenant_select(cls, fallback_tenant_id: UUID = None) -> Select:
-        from api.endpoints.dependencies.tenant_security import get_from_context
 
         result = None
         tenant_context_id = get_from_context("TENANT_ID")
@@ -94,7 +101,9 @@ class TenantScopedModel(BaseModel):
             result = select(cls).where(cls.tenant_id == fallback_tenant_id)
         # couldn't load tenant context for some reason
         else:
-            logger.error("not a http context, no tenant_id, EXECUTING AN UNSAFE QUERY")
+            logger.warning(
+                "not an http context, no fallback_tenant_id, EXECUTING AN UNSAFE QUERY"
+            )
             result = select(cls)
 
         return result
@@ -187,7 +196,7 @@ class Timeline(StatefulModel, table=True):
             .order_by(desc(cls.created_at))
         )
         q_result = await db.execute(q)
-        db_items = q_result.scalars()
+        db_items = q_result.scalars().all()
         return db_items
