+++ services/traction/api/endpoints/dependencies/tenant_context.py
@@ -1,0 +1,13 @@
+from starlette_context import context
+from starlette.exceptions import HTTPException
+from starlette import status
+
+
+def get_from_context(name: str):
+    result = context.get(name)
+    if not result:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Error not authenticated",
+        )
+    return result
