+++ services/traction/api/endpoints/routes/credentials.py
@@ -24,7 +24,7 @@ from api.db.repositories.tenant_workflows import TenantWorkflowsRepository
 
 
 from api.endpoints.dependencies.db import get_db
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 from api.endpoints.models.credentials import (
     CredentialRoleType,
     PresentationRoleType,
