+++ services/traction/api/endpoints/routes/v1/tenant/admin/configuration.py
@@ -10,7 +10,7 @@ from api.endpoints.models.v1.tenant import (
     UpdateTenantConfigurationResponse,
 )
 from api.services.v1 import tenant_configuration_service as service
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 
 router = APIRouter()
 logger = logging.getLogger(__name__)
