+++ services/traction/api/endpoints/routes/v1/tenant/admin/issuer.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter
 
 from starlette import status
 
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 
 from api.endpoints.models.v1.tenant import TenantGetResponse
