+++ services/traction/api/endpoints/routes/v1/tenant/contacts/contacts.py
@@ -6,7 +6,7 @@ from starlette import status
 
 from api.core.config import settings
 from api.endpoints.dependencies.db import get_db
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 from api.endpoints.models.connections import ConnectionRoleType
 
 from api.endpoints.models.v1.contacts import (
