+++ services/traction/api/endpoints/routes/v1/tenant/governance/credential_templates.py
@@ -8,7 +8,7 @@ from starlette.requests import Request
 
 from api.core.config import settings
 from api.endpoints.dependencies.db import get_db
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 from api.endpoints.routes.v1.link_utils import build_list_links
 
 from api.services.v1 import governance_service
