+++ services/traction/api/endpoints/routes/v1/tenant/issuer/credential.py
@@ -7,7 +7,7 @@ from starlette import status
 from starlette.requests import Request
 
 from api.endpoints.dependencies.db import get_db
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 from api.endpoints.models.v1.issuer import (
     IssuerCredentialGetResponse,
     UpdateIssuerCredentialPayload,
