+++ services/traction/api/endpoints/routes/v1/tenant/issuer/credentials.py
@@ -10,7 +10,7 @@ from api.core.config import settings
 from api.endpoints.routes.v1.link_utils import build_list_links
 from api.services.v1 import issuer_service
 
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 from api.endpoints.dependencies.db import get_db
 
 from api.endpoints.models.v1.issuer import (
