+++ services/traction/api/endpoints/routes/v1/tenant/messages/messages.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Request
 from starlette import status
 
 from api.core.config import settings
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 from api.endpoints.models.v1.messages import (
     MessageListResponse,
     MessageListParameters,
