+++ services/traction/api/endpoints/routes/v1/tenant/verifier/verifier_presentation.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter
 
 from api.services.v1 import verifier_service
 
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 
 from api.endpoints.models.v1.verifier import (
     GetVerifierPresentationResponse,
