+++ services/traction/api/endpoints/routes/v1/tenant/verifier/verifier_presentations.py
@@ -6,7 +6,7 @@ from starlette import status
 from api.services.v1 import verifier_service
 from api.core.config import settings
 
-from api.endpoints.dependencies.tenant_security import get_from_context
+from api.endpoints.dependencies.tenant_context import get_from_context
 
 from api.endpoints.models.v1.verifier import (
     VerifierPresentationListResponse,
